refactor: move getKeyStageTitle helper function into formatting file
