+++ src/app/(core)/programmes/[subjectPhaseSlug]/[tab]/Components/UnitSequence/getSavePropsForUnitCard.ts
@@ -1,4 +1,4 @@
-import { getKeyStageTitle } from "./UnitList";
+import { getKeyStageTitle } from "@/utils/curriculum/formatting";
 
 export const getSavePropsForUnitCard = ({
   slug,
