+++ src/utils/curriculum/formatting.ts
@@ -9,6 +9,8 @@ import { sortYears } from "./sorting";
 
 import { Phase } from "@/node-lib/curriculum-api-2023";
 import { DownloadCategory } from "@/node-lib/curriculum-api-2023/fixtures/curriculumPreviousDownloads.fixture";
+import { KeyStageTitleValueType } from "@/browser-lib/avo/Avo";
+import { KeystageSlug } from "@/node-lib/curriculum-api-2023/shared.schema";
 
 export function getYearGroupTitle(
   yearData: YearData,
@@ -27,6 +29,27 @@ export function getYearGroupTitle(
 
 function hasKs(keystages: { slug: string }[], num: number) {
   return keystages.find((k) => k.slug === `ks${num}`);
+}
+
+// Types are loose coming out of the API so we cast to `KeystageSlug` to
+// do our best to map it to the correct title. Fallback to undefined if we can't map it.
+export function getKeyStageTitle(
+  ksSlug: string,
+): KeyStageTitleValueType | undefined {
+  switch (ksSlug as KeystageSlug) {
+    case "ks1":
+      return "Key stage 1";
+    case "ks2":
+      return "Key stage 2";
+    case "ks3":
+      return "Key stage 3";
+    case "ks4":
+      return "Key stage 4";
+    case "early-years-foundation-stage":
+      return "Early Years Foundation stage";
+    default:
+      return undefined;
+  }
 }
 
 export function getPhaseText(
